+++ src/benchmarks/CMakeLists.txt
@@ -3,7 +3,7 @@ cmake_minimum_required(VERSION 3.2 FATAL_ERROR)
 ##################################
 # Benchmark executable
 
-add_executable(end_to_end_bench src/end_to_end_bench.cpp)
+add_executable(end_to_end_bench src/end_to_end_bench.cpp src/bench_utils.cpp src/include/bench_utils.hpp)
 
 target_link_libraries(end_to_end_bench clipper boost cxxopts)
 
@@ -27,6 +27,6 @@ target_link_libraries(rpctest clipper cxxopts)
 ##################################
 # Throughput with Noop Model Testing executable
 
-add_executable(noop_bench src/noop_bench.cpp)
+add_executable(noop_bench src/noop_bench.cpp src/bench_utils.cpp src/include/bench_utils.hpp)
 
 target_link_libraries(noop_bench clipper boost cxxopts)
