+++ src/benchmarks/src/noop_bench.cpp
@@ -1,15 +1,12 @@
 #include <time.h>
 #include <functional>
 #include <iostream>
-#include <memory>
-#include <string>
-#include <utility>
 #include <vector>
 
 #include <boost/thread.hpp>
 #include <cxxopts.hpp>
 
-#include <rapidjson/document.h>
+#include <clipper/app_metrics.hpp>
 #include <clipper/constants.hpp>
 #include <clipper/datatypes.hpp>
 #include <clipper/future.hpp>
@@ -18,52 +15,17 @@
 #include <clipper/query_processor.hpp>
 #include <fstream>
 
+#include "include/bench_utils.hpp"
+
 using namespace clipper;
+using namespace bench_utils;
 
 const std::string DEFAULT_OUTPUT = "-1";
-const std::string TEST_APPLICATION_LABEL = "test";
+const std::string TEST_APPLICATION_LABEL = "throughput_testing_app";
 const std::string NOOP_MODEL_NAME = "bench_noop";
 
 const int RANDOM_VEC_LEN = 784;
 const int RANDOM_VEC_UPPERBOUND = 10;
-
-const std::string CONFIG_KEY_NUM_THREADS = "num_threads";
-const std::string CONFIG_KEY_NUM_BATCHES = "num_batches";
-const std::string CONFIG_KEY_BATCH_SIZE = "batch_size";
-const std::string CONFIG_KEY_BATCH_DELAY = "batch_delay";
-
-
-class AppMetrics {
-  public:
-    explicit AppMetrics(std::string app_name)
-        : app_name_(app_name),
-          latency_(
-                  clipper::metrics::MetricsRegistry::get_metrics().create_histogram(
-                          "app:" + app_name + ":prediction_latency", "microseconds",
-                          4096)),
-          throughput_(
-                  clipper::metrics::MetricsRegistry::get_metrics().create_meter(
-                          "app:" + app_name + ":prediction_throughput")),
-          num_predictions_(
-                  clipper::metrics::MetricsRegistry::get_metrics().create_counter(
-                          "app:" + app_name + ":num_predictions")),
-          default_pred_ratio_(
-                  clipper::metrics::MetricsRegistry::get_metrics()
-                          .create_ratio_counter("app:" + app_name +
-                                                ":default_prediction_ratio")) {}
-      ~AppMetrics() = default;
-    AppMetrics(const AppMetrics&) = default;
-    AppMetrics& operator=(const AppMetrics&) = default;
-
-    AppMetrics(AppMetrics&&) = default;
-    AppMetrics& operator=(AppMetrics&&) = default;
-
-    std::string app_name_;
-    std::shared_ptr<clipper::metrics::Histogram> latency_;
-    std::shared_ptr<clipper::metrics::Meter> throughput_;
-    std::shared_ptr<clipper::metrics::Counter> num_predictions_;
-    std::shared_ptr<clipper::metrics::RatioCounter> default_pred_ratio_;
-};
 
 std::vector<double> gen_random_vector(int length, int upper_bound) {
   std::vector<double> vec(length);
@@ -72,7 +34,6 @@ std::vector<double> gen_random_vector(int length, int upper_bound) {
   }
   return vec;
 }
-
 
 void send_predictions(
     std::unordered_map<std::string, std::string> &config,
@@ -83,10 +44,7 @@ void send_predictions(
   long batch_delay_millis =
           static_cast<long>(std::stoi(config.find(CONFIG_KEY_BATCH_DELAY)->second));
 
-  AppMetrics app_metrics(TEST_APPLICATION_LABEL);
-  std::shared_ptr<metrics::Histogram> msg_latency_hist_;
-
-  // get query vec
+  clipper::app_metrics::AppMetrics app_metrics(TEST_APPLICATION_LABEL);
   std::vector<double> query_vec;
 
   for (int j = 0; j < num_batches; j++) {
@@ -100,7 +58,7 @@ void send_predictions(
                     input,
                     100000,
                     clipper::DefaultOutputSelectionPolicy::get_name(),
-                    {std::make_pair(NOOP_MODEL_NAME, "1")}});
+                    {std::make_pair(NOOP_MODEL_NAME, 1)}});
 
     prediction.then([app_metrics](boost::future<Response> f) {
         Response r = f.get();
@@ -120,44 +78,11 @@ void send_predictions(
   }
 }
 
-std::unordered_map<std::string, std::string> create_config(
-        std::string num_threads, std::string num_batches,
-        std::string batch_size, std::string batch_delay) {
-  std::unordered_map<std::string, std::string> config;
-  config.emplace(CONFIG_KEY_NUM_THREADS, num_threads);
-  config.emplace(CONFIG_KEY_NUM_BATCHES, num_batches);
-  config.emplace(CONFIG_KEY_BATCH_SIZE, batch_size);
-  config.emplace(CONFIG_KEY_BATCH_DELAY, batch_delay);
-  return config;
-};
-
-std::unordered_map<std::string, std::string> get_config_from_prompt() {
-  std::string num_threads;
-  std::string num_batches;
-  std::string batch_size;
-  std::string batch_delay;
-
-  std::cout << "Before proceeding, run bench/noop_bench_setup.sh from clipper's "
-          "root directory."
-            << std::endl;
-  std::cout << "Enter the number of threads of execution: ";
-  std::cin >> num_threads;
-  std::cout
-          << "Enter the number of request batches to be sent by each thread: ";
-  std::cin >> num_batches;
-  std::cout << "Enter the number of requests per batch: ";
-  std::cin >> batch_size;
-  std::cout << "Enter the delay between batches, in milliseconds: ";
-  std::cin >> batch_delay;
-
-  return create_config(num_threads, num_batches, batch_size, batch_delay);
-};
-
 int main(int argc, char *argv[]) {
   cxxopts::Options options("noop_bench",
                            "Clipper noop performance benchmarking");
 
-  std::unordered_map<std::string, std::string> test_config = get_config_from_prompt();
+  std::unordered_map<std::string, std::string> test_config = bench_utils::get_config_from_prompt();
 
   get_config().ready();
   QueryProcessor qp;
