+++ src/frontends/src/query_frontend_main.cpp
@@ -2,6 +2,7 @@
 #include <cxxopts.hpp>
 
 #include <clipper/config.hpp>
+#include <clipper/app_metrics.hpp>
 #include <clipper/constants.hpp>
 #include <clipper/query_processor.hpp>
