+++ src/libclipper/include/clipper/app_metrics.hpp
@@ -1,0 +1,46 @@
+#ifndef CLIPPER_APP_METRICS_H
+#define CLIPPER_APP_METRICS_H
+
+#include <clipper/metrics.hpp>
+
+namespace clipper {
+namespace app_metrics {
+class AppMetrics {
+public:
+    explicit AppMetrics(std::string app_name)
+            : app_name_(app_name),
+              latency_(
+                      clipper::metrics::MetricsRegistry::get_metrics().create_histogram(
+                              "app:" + app_name + ":prediction_latency", "microseconds",
+                              4096)),
+              throughput_(
+                      clipper::metrics::MetricsRegistry::get_metrics().create_meter(
+                              "app:" + app_name + ":prediction_throughput")),
+              num_predictions_(
+                      clipper::metrics::MetricsRegistry::get_metrics().create_counter(
+                              "app:" + app_name + ":num_predictions")),
+              default_pred_ratio_(
+                      clipper::metrics::MetricsRegistry::get_metrics()
+                              .create_ratio_counter("app:" + app_name +
+                                                    ":default_prediction_ratio")) {}
+    ~AppMetrics() = default;
+
+    AppMetrics(const AppMetrics &) = default;
+
+    AppMetrics &operator=(const AppMetrics &) = default;
+
+    AppMetrics(AppMetrics &&) = default;
+
+    AppMetrics &operator=(AppMetrics &&) = default;
+
+    std::string app_name_;
+    std::shared_ptr<clipper::metrics::Histogram> latency_;
+    std::shared_ptr<clipper::metrics::Meter> throughput_;
+    std::shared_ptr<clipper::metrics::Counter> num_predictions_;
+    std::shared_ptr<clipper::metrics::RatioCounter> default_pred_ratio_;
+};
+
+} // namespace app_metrics
+} // namespace clipper
+
+#endif //CLIPPER_APP_METRICS_H
